test: extract `runTsc`
