+++ packages/tsc/tests/build.spec.ts
@@ -1,34 +1,8 @@
 import * as path from 'node:path';
 import { expect, test } from 'vitest';
 import { run } from '..';
+import { runTsc } from './utils';
 
 test(`vue-tsc --build`, () => {
-	expect(runTsc).not.toThrow();
+	expect(() => runTsc('tscBuild')).not.toThrow();
 });
-
-function runTsc() {
-	const originalConsoleLog = process.stdout.write;
-	const originalArgv = process.argv;
-	const originalExit = process.exit;
-	process.stdout.write = () => true;
-	process.argv = [
-		...originalArgv,
-		'--build',
-		path.resolve(__dirname, '../../../test-workspace/tscBuild'),
-		'--pretty',
-		'false',
-	];
-	process.exit = (() => {}) as typeof process.exit;
-	try {
-		const tscPath = require.resolve(
-			`typescript/lib/tsc`,
-			{ paths: [path.resolve(__dirname, '../../../test-workspace')] },
-		);
-		run(tscPath);
-	}
-	finally {
-		process.stdout.write = originalConsoleLog;
-		process.argv = originalArgv;
-		process.exit = originalExit;
-	}
-}
