+++ packages/tsc/tests/utils.ts
@@ -1,0 +1,30 @@
+import * as path from 'node:path';
+import { run } from '..';
+
+export function runTsc(projectName: string) {
+	const consoleOutput: string[] = [];
+	const originalConsoleLog = process.stdout.write;
+	const originalArgv = process.argv;
+	process.stdout.write = output => {
+		consoleOutput.push(String(output).trim());
+		return true;
+	};
+	process.argv = [
+		...originalArgv,
+		'--build',
+		path.resolve(__dirname, `../../../test-workspace/${projectName}`),
+		'--pretty',
+		'false',
+	];
+	try {
+		const tscPath = require.resolve(
+			`typescript/lib/tsc`,
+			{ paths: [path.resolve(__dirname, '../../../test-workspace')] },
+		);
+		run(tscPath);
+	}
+	catch {}
+	process.stdout.write = originalConsoleLog;
+	process.argv = originalArgv;
+	return consoleOutput;
+}
