+++ test-workspace/tscBuild/petiteVue/tsconfig.json
@@ -1,7 +1,7 @@
 {
 	"extends": "../../tsconfig.base.json",
-	"include": ["**/*"],
 	"vueCompilerOptions": {
 		"petiteVueExtensions": [".html"]
-	}
+	},
+	"include": ["**/*"]
 }
