+++ test-workspace/tscBuild/vitepress/tsconfig.json
@@ -1,7 +1,7 @@
 {
 	"extends": "../../tsconfig.base.json",
-	"include": ["**/*"],
 	"vueCompilerOptions": {
 		"vitePressExtensions": [".md"]
-	}
+	},
+	"include": ["**/*"]
 }
